requirements.txt
feedparser==6.0.11
PyYAML==6.0.2
rfeed==1.1.1
sgmllib3k==1.0.0
